requirements.txt
asgiref==3.5.2
Django==4.1.3
django-admin-thumbnails==0.2.6
django-ckeditor==6.5.1
django-extensions==3.2.1
django-filter==22.1
django-htmx==1.13.0
django-js-asset==2.0.0
gunicorn==20.1.0
Pillow==9.3.0
psycopg2==2.9.5
python-decouple==3.6
python-dotenv==0.21.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
whitenoise==6.2.0
